fix: bonus

+++ bonus/includes/libft.hpp
@@ -35,7 +35,7 @@ namespace ft
     size_t pow(size_t root, size_t square);
     long long int abs(long long int num);
     std::string to_string(long long int n);
-    int stoi(std::string str, int base = 10);
+    int stoi(std::string str, size_t base = 10);
     std::string itos(std::string number, size_t from, size_t to);
 
     /* C++ LIBFT */
